+++ src/main/java/com/jlgs/howmuchah/entity/GroupMemberId.java
@@ -1,0 +1,19 @@
+package com.jlgs.howmuchah.entity;
+
+import jakarta.persistence.Embeddable;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.io.Serializable;
+import java.util.UUID;
+
+@Embeddable
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class GroupMemberId implements Serializable {
+
+    private UUID groupId;
+    private UUID userId;
+}